refactor: Move run_job and run_jobs functions to job module and simplify job creation in main.rs
